requirements.txt
google-api-python-client>=2.129.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
pdf2image>=1.17.0
openai>=1.23.6
pydantic>=2.0.0
pillow>=10.3.0
pandas>=2.2.2
numpy>=
gspread>=6.1.2
